+++ CHANGES.md
@@ -3,4 +3,10 @@
 * Update version of angular to 1.2.3
 * List dependencies on angular and pouchdb latest explicitly.
 
+# v. 0.1.3
+
+* Add a `pouch-repeat` directive, to traverse all elements in a collection.
+* Add support for sorting in `pouch-repeat`.
+
+